Refactor: useDailyData 훅으로 분리 #79
