+++ src/pages/Todo/Todo.jsx
@@ -1,4 +1,4 @@
-import ShowDate from "components/ShowDate/ShowDate"; 
+import ShowDate from "components/ShowDate/ShowDate";
 import ImgSave from "components/ImgSave/ImgSave";
 import TodoCategoryBtn from "components/TodoCategoryBtn/TodoCategoryBtn";
 import TodoLists from "components/TodoLists/TodoLists";
@@ -40,7 +40,7 @@ const Todo = () => {
 
   useEffect(() => {
     if (categories.length > 0) {
-      categories.forEach(category => {
+      categories.forEach((category) => {
         fetchTodos(date, category.categoryId);
       });
     }
@@ -51,9 +51,9 @@ const Todo = () => {
       console.error("🚨 Error: date 값이 없습니다!", selectedDate);
       return;
     }
-  
+
     console.log("📅 fetchTodos 실행 - 전달된 date 값:", selectedDate);
-  
+
     setIsLoading(true);
     try {
       const accessToken = localStorage.getItem("accessToken");
@@ -67,23 +67,21 @@ const Todo = () => {
           Authorization: `Bearer ${accessToken}`,
         },
       });
-  
+
       const todoList = response.data.result.todoList;
-      setTodos(prevTodos => ({
+      setTodos((prevTodos) => ({
         ...prevTodos,
-        [categories.find(cat => cat.categoryId === categoryId).categoryName]: todoList.map(todo => (
-          {
-          
+        [categories.find((cat) => cat.categoryId === categoryId).categoryName]: todoList.map((todo) => ({
           name: todo.todoName,
           categoryId: todo.categoryId,
           status: todo.status,
-          todoId: todo.todoId, 
-          startTime: todo.startTime, 
-          endTime: todo.endTime, 
+          todoId: todo.todoId,
+          startTime: todo.startTime,
+          endTime: todo.endTime,
           formattedDuration: todo.formattedDuration,
         })),
       }));
-  
+
       console.log("📌 response.data.result.todoList:", todoList);
     } catch (error) {
       console.error("🚨 TODO 데이터 가져오기 실패:", error);
@@ -92,12 +90,10 @@ const Todo = () => {
     }
   };
 
-
-
   const handleAddTodoDirectly = (categoryName) => {
-    const category = categories.find(cat => cat.categoryName === categoryName);
+    const category = categories.find((cat) => cat.categoryName === categoryName);
     if (!category) return;
-  
+
     setTodos({
       ...todos,
       [categoryName]: [
@@ -108,67 +104,73 @@ const Todo = () => {
           date: new Date().toISOString().split("T")[0], // 오늘 날짜 기본값
           startTime: "00:00", // 기본값
           endTime: "00:00", // 기본값
-          status: 2, 
+          status: 2,
           todoId: null, // POST 후 서버에서 받음
         },
       ],
     });
   };
-  
+
   const handleTodoTextChange = async (categoryName, index, newText, status, todoId) => {
-    const category = categories.find(cat => cat.categoryName === categoryName);
+    const category = categories.find((cat) => cat.categoryName === categoryName);
     if (!category) return;
-  
+
     const todo = todos[categoryName][index];
-  
+
     if (!todo) {
       console.error("🚨 수정하려는 Todo를 찾을 수 없습니다!");
       return;
     }
-  
+
     if (todo.name === newText) {
       console.warn("🚨 동일한 데이터 중복 요청 방지:", newText);
       return;
     }
-  
+
     try {
       const accessToken = localStorage.getItem("accessToken");
-  
+
       if (!todoId) {
         // 🟢 POST 요청 (새로운 Todo 생성)
-        await axiosInstance.post(`/api/v0/todos/`, {
-          todoName: newText || "새로운 할 일",
-          date: date || new Date().toISOString().split("T")[0],
-          categoryId: category.categoryId,
-          startTime: "00:00",
-          endTime: "00:00",
-          status: status || 1,
-        }, {
-          headers: { Authorization: `Bearer ${accessToken}` },
-        });
-  
+        await axiosInstance.post(
+          `/api/v0/todos/`,
+          {
+            todoName: newText || "새로운 할 일",
+            date: date || new Date().toISOString().split("T")[0],
+            categoryId: category.categoryId,
+            startTime: "00:00",
+            endTime: "00:00",
+            status: status || 1,
+          },
+          {
+            headers: { Authorization: `Bearer ${accessToken}` },
+          }
+        );
+
         console.log("✅ 새 Todo 생성 성공. ID 없음 -> GET 요청 실행");
-  
+
         // 🔥 새로 생성된 todoId를 GET 요청으로 가져오기
         await fetchTodos(date, category.categoryId);
       } else {
         // 🟢 PATCH 요청 (기존 Todo 수정)
-        const response = await axiosInstance.patch(`/api/v0/todos/${todoId}`, {
-          todoName: newText,
-          date,
-        }, {
-          headers: { Authorization: `Bearer ${accessToken}` },
-        });
-  
+        const response = await axiosInstance.patch(
+          `/api/v0/todos/${todoId}`,
+          {
+            todoName: newText,
+            date,
+          },
+          {
+            headers: { Authorization: `Bearer ${accessToken}` },
+          }
+        );
+
         console.log("✅ Todo 수정 성공:", response.data);
         await fetchTodos(date, category.categoryId);
-  
+
         // 🔥 기존 Todo 상태 업데이트
-        setTodos(prevTodos => ({
+        setTodos((prevTodos) => ({
           ...prevTodos,
-          [categoryName]: prevTodos[categoryName].map((t, i) =>
-            i === index ? { ...t, name: newText, status } : t
-          ),
+          [categoryName]: prevTodos[categoryName].map((t, i) => (i === index ? { ...t, name: newText, status } : t)),
         }));
       }
     } catch (error) {
@@ -183,53 +185,54 @@ const Todo = () => {
     }
     try {
       const accessToken = localStorage.getItem("accessToken");
-  
+
       // 🛠️ 서버에서 삭제 요청
       await axiosInstance.delete(`/api/v0/todos/${todoId}`, {
         headers: { Authorization: `Bearer ${accessToken}` },
       });
-  
+
       console.log("✅ Todo 삭제 성공:", todoId);
-  
+
       // 🔥 UI에서도 즉시 삭제
-      setTodos(prevTodos => {
+      setTodos((prevTodos) => {
         const updatedTodos = { ...prevTodos };
         for (const categoryName in updatedTodos) {
-          updatedTodos[categoryName] = updatedTodos[categoryName].filter(todo => todo.todoId !== todoId);
+          updatedTodos[categoryName] = updatedTodos[categoryName].filter((todo) => todo.todoId !== todoId);
         }
         return updatedTodos;
       });
-  
     } catch (error) {
       console.error("🚨 Todo 삭제 실패:", error);
     }
   };
-  
 
   const handleTimeUpdate = async (todoId, startTime, endTime) => {
     try {
       const accessToken = localStorage.getItem("accessToken");
-      const response = await axiosInstance.patch(`/api/v0/todos/${todoId}`, {
-        startTime,
-        endTime,
-        date,
-      }, {
-        headers: { Authorization: `Bearer ${accessToken}` },
-      });
-  
+      const response = await axiosInstance.patch(
+        `/api/v0/todos/${todoId}`,
+        {
+          startTime,
+          endTime,
+          date,
+        },
+        {
+          headers: { Authorization: `Bearer ${accessToken}` },
+        }
+      );
+
       console.log("✅ Time update 성공:", response.data);
-  
+
       // 프론트엔드 상태 업데이트
-      setTodos(prevTodos => {
+      setTodos((prevTodos) => {
         const updatedTodos = { ...prevTodos };
         for (const categoryName in updatedTodos) {
-          updatedTodos[categoryName] = updatedTodos[categoryName].map(todo =>
+          updatedTodos[categoryName] = updatedTodos[categoryName].map((todo) =>
             todo.todoId === todoId ? { ...todo, startTime, endTime } : todo
           );
         }
         return updatedTodos;
       });
-  
     } catch (error) {
       console.error("🚨 Time update 실패:", error);
     }
@@ -237,7 +240,7 @@ const Todo = () => {
 
   return (
     <S.DailyContainer className="ImgContainer">
-      <ShowDate date={date}  onDateChange={setDate} />
+      <ShowDate date={date} onDateChange={setDate} />
       {isLoading && <p>Loading...</p>}
       <S.TodoCategoryContainer>
         {categories.map((category) => (
@@ -250,9 +253,15 @@ const Todo = () => {
             <TodoLists
               todos={todos[category.categoryName]}
               onTodoTextChange={(index, newText, status) =>
-                handleTodoTextChange(category.categoryName, index, newText, status, todos[category.categoryName][index]?.todoId)
+                handleTodoTextChange(
+                  category.categoryName,
+                  index,
+                  newText,
+                  status,
+                  todos[category.categoryName][index]?.todoId
+                )
               }
-              onDeleteTodo={deleteTodo}  
+              onDeleteTodo={deleteTodo}
               category={category}
               date={date}
               onTimeUpdate={handleTimeUpdate}
